+++ Models/ViewModels/AdminOrderViewModels.cs
@@ -7,16 +7,16 @@ namespace ValiModern.Models.ViewModels
 {
     public class OrderListItemVM
     {
-  public int Id { get; set; }
+        public int Id { get; set; }
         public int UserId { get; set; }
         public string UserName { get; set; }
         public string UserEmail { get; set; }
         public DateTime OrderDate { get; set; }
-  public string Status { get; set; }
-   public decimal TotalAmount { get; set; }
+        public string Status { get; set; }
+        public long TotalAmount { get; set; }
         public string Phone { get; set; }
-    public string ShippingAddress { get; set; }
-  public int ItemCount { get; set; }
+        public string ShippingAddress { get; set; }
+        public int ItemCount { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
@@ -26,12 +26,12 @@ namespace ValiModern.Models.ViewModels
         public int Id { get; set; }
         public int UserId { get; set; }
         public string UserName { get; set; }
-      public string UserEmail { get; set; }
+        public string UserEmail { get; set; }
         public string UserPhone { get; set; }
         public DateTime OrderDate { get; set; }
-      public string Status { get; set; }
-  public decimal TotalAmount { get; set; }
- public string Phone { get; set; }
+        public string Status { get; set; }
+        public long TotalAmount { get; set; }
+        public string Phone { get; set; }
         public string ShippingAddress { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
@@ -39,29 +39,29 @@ namespace ValiModern.Models.ViewModels
         // Order items
         public List<OrderDetailItemVM> Items { get; set; } = new List<OrderDetailItemVM>();
 
-     // Payment info
-     public List<PaymentVM> Payments { get; set; } = new List<PaymentVM>();
+        // Payment info
+        public List<PaymentVM> Payments { get; set; } = new List<PaymentVM>();
     }
 
     public class OrderDetailItemVM
     {
-      public int Id { get; set; }
-  public int ProductId { get; set; }
+        public int Id { get; set; }
+        public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductImageUrl { get; set; }
-   public int Quantity { get; set; }
-        public decimal Price { get; set; }
-  public decimal Subtotal { get; set; }
+        public int Quantity { get; set; }
+        public int Price { get; set; }
+        public long Subtotal { get; set; }
         public string ColorName { get; set; }
         public string ColorCode { get; set; }
-   public string SizeName { get; set; }
+        public string SizeName { get; set; }
         public DateTime CreatedAt { get; set; }
     }
 
     public class PaymentVM
     {
- public int Id { get; set; }
-        public decimal Amount { get; set; }
+        public int Id { get; set; }
+        public long Amount { get; set; }
         public string PaymentMethod { get; set; }
         public string Status { get; set; }
         public string TransactionId { get; set; }
