+++ Models/ViewModels/CartViewModels.cs
@@ -11,7 +11,7 @@ namespace ValiModern.Models.ViewModels
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductImageUrl { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public int Quantity { get; set; }
         public int? ColorId { get; set; }
         public string ColorName { get; set; }
@@ -19,14 +19,14 @@ namespace ValiModern.Models.ViewModels
         public int? SizeId { get; set; }
         public string SizeName { get; set; }
 
-        public decimal Subtotal => Price * Quantity;
+        public long Subtotal => (long)Price * Quantity;
     }
 
     // Cart view model
     public class CartVM
     {
         public List<CartItem> Items { get; set; } = new List<CartItem>();
-        public decimal TotalAmount => Items.Sum(i => i.Subtotal);
+        public long TotalAmount => Items.Sum(i => i.Subtotal);
         public int TotalItems => Items.Sum(i => i.Quantity);
     }
 
@@ -34,7 +34,7 @@ namespace ValiModern.Models.ViewModels
     public class CheckoutVM
     {
         public List<CartItem> Items { get; set; } = new List<CartItem>();
-        public decimal TotalAmount => Items.Sum(i => i.Subtotal);
+        public long TotalAmount => Items.Sum(i => i.Subtotal);
 
         // Shipping information
         public string FullName { get; set; }
@@ -52,7 +52,7 @@ namespace ValiModern.Models.ViewModels
     {
         public int OrderId { get; set; }
         public string OrderCode { get; set; }
-        public decimal TotalAmount { get; set; }
+        public long TotalAmount { get; set; }
         public string PaymentMethod { get; set; }
         public string Status { get; set; }
         public DateTime OrderDate { get; set; }
