+++ Models/ViewModels/HomeViewModels.cs
@@ -7,15 +7,15 @@ namespace ValiModern.Models.ViewModels
     {
         public int id { get; set; }
         public string name { get; set; }
-        public decimal price { get; set; }
-        public decimal original_price { get; set; }
+        public int price { get; set; }
+        public int original_price { get; set; }
         public string image_url { get; set; }
         public string description { get; set; }
         public int sold { get; set; }
         public string brand_name { get; set; }
         public string category_name { get; set; }
         public bool HasDiscount => original_price > price && original_price > 0;
-        public int DiscountPercent => HasDiscount ? (int)((original_price - price) / original_price * 100) : 0;
+        public int DiscountPercent => HasDiscount ? (int)((original_price - price) / (double)original_price * 100) : 0;
     }
 
     public class CategoryBlockVM
@@ -43,8 +43,8 @@ namespace ValiModern.Models.ViewModels
         public string CurrentBrand { get; set; }
         public string CurrentSort { get; set; }
         public string SearchQuery { get; set; }
-        public decimal? MinPrice { get; set; }
-        public decimal? MaxPrice { get; set; }
+        public int? MinPrice { get; set; }
+        public int? MaxPrice { get; set; }
         public List<string> SelectedColorNames { get; set; }
         public List<string> SelectedSizeNames { get; set; }
         public int CurrentPage { get; set; }
